// src/types/marketplace-types.d.ts
export type RepoTopic = "navify-extensions" | "navify-themes" | "navify-apps";

export type TabType = "Extensions" | "Themes" | "Snippets" | "Apps" | "Installed";

export type ResetCategory = "extensions" | "snippets" | "theme";

export type CardType = "extension" | "theme" | "snippet" | "app";

export type RepoType = "extension" | "theme" | "app";

export type Author = {
  name: string;
  url: string;
};

// From snippets.json
export type Snippet = {
  title: string;
  description: string;
  code: string;

  // preview is used in the JSON to stay consistent with the other manifest formats
  preview?: string;
  // it's parsed into an imageURL because that's what the Card component uses
  imageURL?: string;

  // TODO: clean this up somehow
  // It complains bitterly in Card.tsx
  // if I don't have all the same properties as CardItem
  manifest: undefined;
  subtitle: undefined;
  authors: undefined;
  user: undefined;
  repo: undefined;
  branch: undefined;
  archived: undefined;
  extensionURL: undefined;
  readmeURL: undefined;
  stars: undefined;
  tags: undefined;
  cssURL: undefined;
  schemesURL: undefined;
  include: undefined;
  lastUpdated: undefined;
  created: undefined;
};

// From `fetchExtensionManifest()` and `fetchThemeManifest()`
export type Manifest = {
  name: string;
  description: string;
  main: string;
  authors: Author[];
  preview: string;
  readme: string;

  // TODO: split these into different types?
  tags?: string[];
  usercss?: string;
  schemes?: string[];
  include?: string[];
};

// From fetchExtensionManifest
// (Card.props.item)
export type CardItem = {
  manifest: Manifest;
  title: string;
  subtitle: string;
  authors: Author[];
  user: string;
  repo: string;
  branch: string;
  archived: boolean;
  imageURL: string;
  extensionURL: string;
  readmeURL: string;
  stars: number;
  tags: string[];
  lastUpdated: string;
  created: string;
  name: string;
  stargazers_count: number;
  // For themes only
  cssURL?: string;
  schemesURL?: string;
  include?: string[];

  // TODO: clean this up somehow
  // It complains bitterly in Card.tsx
  // if I don't have all the same properties as CardItem
  code: undefined;
  description: undefined;
};